Sentry NSError in Swift
