+++ Sources/Sentry/Public/SentryWithoutUIKit.h
@@ -31,7 +31,6 @@ FOUNDATION_EXPORT const unsigned char SentryVersionString[];
 #    import <SentryWithoutUIKit/SentryMechanism.h>
 #    import <SentryWithoutUIKit/SentryMechanismMeta.h>
 #    import <SentryWithoutUIKit/SentryMessage.h>
-#    import <SentryWithoutUIKit/SentryNSError.h>
 #    import <SentryWithoutUIKit/SentryOptions.h>
 #    import <SentryWithoutUIKit/SentryReplayApi.h>
 #    import <SentryWithoutUIKit/SentryRequest.h>
