+++ Sources/Sentry/SentryMechanismMeta.m
@@ -1,6 +1,6 @@
 #import "SentryMechanismMeta.h"
 #import "SentryNSDictionarySanitize.h"
-#import "SentryNSError.h"
+#import "SentrySwift.h"
 
 NS_ASSUME_NONNULL_BEGIN
