+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -152,7 +152,6 @@
 #import "SentryNSDataUtils.h"
 #import "SentryNSDictionarySanitize+Tests.h"
 #import "SentryNSDictionarySanitize.h"
-#import "SentryNSError.h"
 #import "SentryNSNotificationCenterWrapper.h"
 #import "SentryNSProcessInfoWrapper.h"
 #import "SentryNSTimerFactory.h"
